refactor(key-value-operations): rename folder and update references
